use size_t instead of UL to appease windows

+++ mt-kahypar/partition/refinement/deterministic/deterministic_label_propagation.cpp
@@ -350,11 +350,11 @@ namespace mt_kahypar {
 
       HypernodeWeight  budget_p1 = context.partition.max_part_weights[p1] - phg.partWeight(p1),
                        budget_p2 = context.partition.max_part_weights[p2] - phg.partWeight(p2);
-      HypernodeWeight  lb_p1 = -(budget_p1 /std::max(1UL, involvements[p1])),
-                       ub_p2 = budget_p2 / std::max(1UL, involvements[p2]);
+      HypernodeWeight  lb_p1 = -(budget_p1 /std::max(size_t(1), involvements[p1])),
+                       ub_p2 = budget_p2 / std::max(size_t(1), involvements[p2]);
 
-      size_t  p1_begin = positions[index(p1, p2)], p1_end = positions[index(p1, p2) + 1],
-              p2_begin = positions[index(p2, p1)], p2_end = positions[index(p2, p1) + 1];
+      size_t p1_begin = positions[index(p1, p2)], p1_end = positions[index(p1, p2) + 1],
+             p2_begin = positions[index(p2, p1)], p2_end = positions[index(p2, p1) + 1];
 
       auto best_prefix = findBestPrefixesRecursive(p1_begin, p1_end, p2_begin, p2_end,
                                                    p1_begin - 1, p2_begin - 1, lb_p1, ub_p2);
